backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.26.0
pydantic==2.6.4
pydantic-settings==2.2.1
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
better-auth==0.0.1b11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
qdrant-client==1.10.0
openai==1.12.0
tiktoken==0.6.0
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.21.1
httpx==0.27.0
pytest-cov==4.1.0
pydantic-factories==1.17.1
factory-boy==3.3.0